% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADtest.R
\name{ADtest}
\alias{ADtest}
\title{ADtest: A function to compare two estimated densities.}
\usage{
ADtest(x1, x2)
}
\arguments{
\item{x1}{A vector of values. Defaults to NULL.}

\item{x2}{A vector of values.  Defaults to NULL.}
}
\value{
A list object, containing the following list elements.

ad$pvalue A pvalue indicating the confidence that the data are the same distribution.

ad$statistics The AD test statistic for comparing the two data sets.
}
\description{
This function takes two vectors of values and compares them, to determine
if they come from the same distribution.
}
\examples{
## Define ctrl object.
x <- rnorm(1000,0,1)
myProbHist(x)

}
\author{
Jennifer Starling
}
\keyword{AD}
\keyword{Anderson-Darling}
\keyword{test}

