% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiFCritVal.R
\name{ChiFCritVal}
\alias{ChiFCritVal}
\title{ChiFCritVal(conf,n,p) A critical value function for Z and T distributions.}
\usage{
ChiFCritVal(conf = c(0.99, 0.98, 0.975, 0.95, 0.9), n = 1000, p = 1,
  n2 = 1000, p2 = 1)
}
\arguments{
\item{conf}{A vector of confidence levels (1-alpha) to calculate critical values.
Defaults to c(.99, .985, .975, .95, .9)}

\item{n}{A sample size for calculating degrees of freedom (df1 for F).  Defaults to 1000.}

\item{p}{The number of parameters being estimated for degrees of freedom (df1 for F).  Defaults to 1.}

\item{n2}{A sample size for calculating degrees of freedom (df2 for F).  Defaults to 1000.}

\item{p2}{The number of parameters being estimated for degrees of freedom (df2 for F).  Defaults to 1.}
}
\value{
Returns a list object x containing the following values.

x$n The sample size input by the user.

x$p The number of estimated parameters input by the user.

x$n2 The second sample size input by the user.

x$p2 The second number of estimated parameters input by the user.

x$twoSided The table of 2-sided critical values for the chi-squred and F dists.

x$oneSided The table of 1-sided critical values (right tail) for the chi-squred and F dists.
}
\description{
This function returns two tables of critical values (one-sided and two-sided)
for the chi-squared and F (central) distributions.
}
\examples{
## Input specific parameters.
ChiFCritVal(conf=c(.99,.95), n=1000, p=1)
## Run with default parameters.
ChiFCritVal()

}
\author{
Jennifer Starling
}
\keyword{critical}
\keyword{hypothesis}
\keyword{testing}
\keyword{values}

