% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossProduct3D.R
\name{CrossProduct3D}
\alias{CrossProduct3D}
\title{CrossProduct3D(a,b,i)}
\usage{
CrossProduct3D(x, y, i = 1:3)
}
\arguments{
\item{i}{An index value of 1, 2, or 3.}

\item{a}{A vector of length 3.}

\item{b}{A vector of length 3.}
}
\value{
A vector (length 3) containing the cross-product of the two input vectors.
}
\description{
This function calculates the cross product of two vectors, each of 
length 3.  Specifying i = 1, 2 or 3 returns a single element of the cross-product
vector, based on the i index.  Leaving i blank returns the entire cross-product vector.
}
\examples{
a <- c(2,3,4)
b <- c(5,6,7)
x <- CrossProduct3D(a,b)

}
\author{
Jennifer Starling
}
\keyword{}

