% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreesToRadians.R
\name{degreesToRadians}
\alias{degreesToRadians}
\title{degreesToRadians(degrees, minutes)}
\usage{
degreesToRadians(degrees = 45, minutes = 0)
}
\arguments{
\item{degrees}{A degrees value.  Defaults to 45.}

\item{minutes}{A minutes value.  Defaults to NULL.}
}
\value{
radians The radians equivalent of the input degrees value.
}
\description{
This function converts degrees to radians.
}
\examples{
radiansToDegrees(45,0)	##Example entering degrees and minutes.
radiansToDegrees()		##Defaults to 45 degrees, 0 minutes.
radiansToDegrees(45)  	##No need to input a minutes value.

}
\author{
Jennifer Starling
}
\keyword{trigonometry}

