% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMatrix.R
\name{genMatrix}
\alias{genMatrix}
\title{genMatrix(type,n=5)}
\usage{
genMatrix(type = "identity", n = 5)
}
\arguments{
\item{type}{The type of the example matrix to generate.  Required field; default is 'identity'.
'zero' to generate a zero matrix.
 'identity' to generate an identity matrix.
 'ut' to generate an upper triangular matrix.
 'uut' to generate a unit upper triangular matrix.
 'ust' to generate a strictly upper triangular matrix.
 'lt' to generate a lower triangular matrix.
 'lut' to generate a unit lower triangular matrix.
 'lst' to generate a strictly lower triangular matrix.}

\item{n}{The number of rows and colums of the square matrix (nxn).  Default is n=5.}
}
\value{
mat A matrix of the type and dimension specified.
}
\description{
This function generates examples of various square matrix types.
}
\examples{
genMatrix('identity',5)

}
\author{
Jennifer Starling
}
\keyword{matrix}

