% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myProbHist.R
\name{myProbHist}
\alias{myProbHist}
\title{myProbHist: A probability histogram function.}
\usage{
myProbHist(x)
}
\arguments{
\item{x}{A vector of values. Defaults to NULL.}
}
\value{
Plots a probability histogram of the x data.
}
\description{
This function plots the probability histogram for the input vector of data (x).
The y-axis is scaled from 0 to 1, to reflect the probability of each bin.
}
\examples{
## Define ctrl object.
x <- rnorm(1000,0,1)
myProbHist(x)

}
\author{
Jennifer Starling
}
\keyword{histogram}
\keyword{probability}

