% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeKDE.R
\name{safeKDE}
\alias{safeKDE}
\title{safeKDE(x,classes,kdeClass)}
\usage{
safeKDE(x, classes = NULL, kdeClass = NULL)
}
\arguments{
\item{x}{A vector of values for which a KDE will be created (either for the entire x vector,
or for just the specified kdeClass, depending on inputs.)}

\item{classes}{A vector of categorical class values for vector x.  Length must equal length(x).}

\item{kdeClass}{The class within classes for which a kde should be created.  If classes and
kdeclass are not populated, a kde is created for the entire x vector.}
}
\value{
ks A list object with three components.

ks$eval.points The x values for the kde estimate.

ks$estimate The y values for the kde estimate.

ks$h The bandwidth for the kde estimate.
}
\description{
Function purpose is to create kernel density estimates for a given vector of data.
Function handles errors seen when the data has a variance of zero.  A vector of 
categorical 'group' levels can be provided (grps input), along with a level
}
\examples{
##data is a data frame containing a column named 'class'.
p <- classProps(data)

}
\author{
Jennifer Starling
}
\keyword{density}
\keyword{estimation}
\keyword{kde}
\keyword{kernel}

