% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_l1.R, R/soft.thresholding.R
\name{soft.thresholding}
\alias{soft.thresholding}
\title{Proximal operator of the scaled L1 norm.}
\usage{
soft.thresholding(x, lambda)

soft.thresholding(x, lambda)
}
\arguments{
\item{x}{The input vector}

\item{t}{The step size}

\item{opts}{List of parameters, which must include:
\itemize{\code{opts$lambda}, the scaling factor of the L1 norm.}}

\item{x}{The input vector}

\item{t}{The step size}

\item{opts}{List of parameters, which must include:
\itemize{\code{opts$lambda}, the scaling factor of the L1 norm.}}
}
\value{
The proximal of \eqn{h} at {x} with step size \eqn{t}, given by
  \deqn{prox_h(x,t) = argmin_u [ t h(u) + 1/2 || x - u ||^2 ]}.

The proximal of \eqn{h} at {x} with step size \eqn{t}, given by
  \deqn{prox_h(x,t) = argmin_u [ t h(u) + 1/2 || x - u ||^2 ]}.
}
\description{
Computes the proximal operator of the L1 norm: \deqn{h(x) = \lambda ||x||_1
,} where \eqn{\lambda} is a scaling factor.

Computes the proximal operator of the L1 norm: \deqn{h(x) = \lambda ||x||_1
,} where \eqn{\lambda} is a scaling factor.
}
\examples{
prox.l1(c(1,3,-2), 1.5, list(lambda=1))
prox.l1(c(1,3,-2), 1.5, list(lambda=1))
}

