% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecAngle.R
\name{vecAngle}
\alias{vecAngle}
\title{vecAngle(x,y)}
\usage{
vecAngle(x, y)
}
\arguments{
\item{x}{A vector.  Defaults to NULL.}

\item{y}{A vector.  Defaults to NULL.}
}
\value{
A scalar, the angle between the two vectors.
}
\description{
This function calculates the angle in degrees between two vectors.
}
\examples{
x <- c(1,3,5,7,9)
y <- c(2,4,6,8,10)
vecAngle(x,y)

}
\author{
Jennifer Starling
}
\keyword{}

