% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSizeCalc.R, R/vecNorm.R
\name{vecNorm}
\alias{vecNorm}
\title{vecNorm(x)}
\usage{
vecNorm(x)

vecNorm(x)
}
\arguments{
\item{x}{A vector.  Defaults to NULL.}

\item{x}{A vector.  Defaults to NULL.}
}
\value{
A scalar, the norm of vector x.

A scalar, the norm of vector x.
}
\description{
This function calculates the norm (magnitude, Euclidian length) of a vector x.

This function calculates the norm (magnitude, Euclidian length) of a vector x.
}
\examples{
x <- c(1,3,5,7,9)
vecNorm(x)

x <- c(1,3,5,7,9)
vecNorm(x)

}
\author{
Jennifer Starling

Jennifer Starling
}
\keyword{}

