% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMatrix.R
\name{genMatrix}
\alias{genMatrix}
\title{genMatrix(type,n=5)}
\usage{
genMatrix(type = "identity", n = 5)
}
\arguments{
\item{type}{The type of the example matrix to generate.  Required field; default is 'identity'.}

\item{n}{The number of rows and colums of the square matrix (nxn).  Default is n=5.}

\item{type='zero'}{to generate a zero matrix.}

\item{type='identity'}{to generate an identity matrix.}

\item{type='ut'}{to generate an upper triangular matrix.}

\item{type='uut'}{to generate a unit upper triangular matrix.}

\item{type='ust'}{to generate a strictly upper triangular matrix.}

\item{type='lt'}{to generate a lower triangular matrix.}

\item{type='lut'}{to generate a unit lower triangular matrix.}

\item{type='lst'}{to generate a strictly lower triangular matrix.}
}
\value{
mat A matrix of the type and dimension specified.
}
\description{
This function generates examples of various square matrix types.
}
\examples{
genMatrix('identity',5)

}
\author{
Jennifer Starling
}
\keyword{matrix}

